requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckedit==1.0
django-ckeditor==6.7.0
django-environ==0.11.2
django-js-asset==2.1.0
django-model-utils==4.3.1
easy-thumbnails==2.8.5
Pillow==10.1.0
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2023.3
Unipath==1.1